requirements.txt
Flask==2.0.1
heroku3==5.1.4
Jinja2==3.0.1
Pillow==8.3.2
pylint==2.8.2
python-dotenv==0.19.0
requests==2.26.0
stripe==2.60.0
gunicorn==20.1.0
dnspython==1.16.0 
python-binance==1.0.15
pandas==1.3.3
python-dateutil==2.8.2
